implemented CRUD operations for all entities

+++ src/main/java/com/softserve/firstdemo/service/CityService.java
@@ -2,17 +2,19 @@ package com.softserve.firstdemo.service;
 
 import com.softserve.firstdemo.dao.CityDao;
 import com.softserve.firstdemo.entity.City;
+import com.softserve.firstdemo.service.validation.StringValidator;
 
 import java.util.List;
 
 public class CityService {
 
-
     private CityDao cityDao = new CityDao();
+    private StringValidator stringValidator = new StringValidator();
 
     public void addCity(String name) {
-
-        City city = new City(name);
+        City city = new City();
+        stringValidator.validateString(name);
+        city.setName(name);
         cityDao.create(city);
     }
 
@@ -21,22 +23,23 @@ public class CityService {
     }
 
     public City findById(int id) {
-        City city = cityDao.readById(id);
-        return city;
+        return cityDao.readById(id);
     }
 
     public City findByName(String name) {
-        City city = cityDao.readCityByName(name);
-        return city;
+        return cityDao.readByName(name);
     }
 
     public void editCity(int id, String name) {
         City city = cityDao.readById(id);
+        stringValidator.validateString(name);
         city.setName(name);
         cityDao.update(city);
     }
 
     public void deleteCity(int id) {
-        cityDao.delete(id);
+        if (id != 0) {
+            cityDao.delete(id);
+        }
     }
 }
